backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# Auth & Security
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
slowapi==0.1.9

# LangChain & LangGraph
langchain==0.3.14
langchain-openai==0.3.0
langchain-core==0.3.30
langgraph==0.2.62

# Job Queue
celery[redis]==5.4.0
redis==5.2.1

# Document Parsing
pymupdf==1.25.1
python-docx==1.1.2

# HTTP Client
httpx==0.28.1

# Configuration
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Utilities
tenacity==9.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
aiosqlite==0.20.0       # async SQLite for in-memory test DB

# Email validation (required by Pydantic EmailStr)
email-validator==2.2.0
